+++ kstars/ekos/auxiliary/opticaltrainmanager.h
@@ -62,8 +62,8 @@ class OpticalTrainManager : public QDialog, public Ui::OpticalTrain
         void setProfile(const QSharedPointer<ProfileInfo> &profile);
         void checkOpticalTrains();
 
-        bool exists(uint8_t id) const;
-        const QVariantMap getOpticalTrain(uint8_t id) const;
+        bool exists(uint32_t id) const;
+        const QVariantMap getOpticalTrain(uint32_t id) const;
         const QVariantMap getOpticalTrain(const QString &name) const;
         const QList<QVariantMap> &getOpticalTrains() const
         {
